// client/src/redux/actionTypes.ts
export type ActionType = 
    "GET_COMPANIES" | "GET_COMPANIES_PASSED" | "GET_COMPANIES_FAILED"
    | "ADD_COMPANY" | "ADD_COMPANY_PASSED" | "ADD_COMPANY_FAILED"
    | "EDIT_COMPANY" | "EDIT_COMPANY_PASSED" | "EDIT_COMPANY_FAILED"
    | "DELETE_COMPANY" | "DELETE_COMPANY_PASSED" | "DELETE_COMPANY_FAILED"

    | "ADD_JOB" | "ADD_JOB_PASSED" | "ADD_JOB_FAILED"
    | "EDIT_JOB" | "EDIT_JOB_PASSED" | "EDIT_JOB_FAILED"
    | "DELETE_JOB" | "DELETE_JOB_PASSED" | "DELETE_JOB_FAILED";